+++ src/types/me.rs
@@ -66,6 +66,8 @@ mod tests {
                 last_name: None,
                 username: Some("SomethingSomethingBot".to_owned()),
                 language_code: None,
+                is_premium: false,
+                added_to_attachment_menu: false,
             },
             can_join_groups: false,
             can_read_all_group_messages: false,