+++ src/types/sticker.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::types::{MaskPosition, PhotoSize};
+use crate::types::{File, MaskPosition, PhotoSize};
 
 /// This object represents a sticker.
 ///
@@ -40,6 +40,9 @@ pub struct Sticker {
 
     /// Name of the sticker set to which the sticker belongs.
     pub set_name: Option<String>,
+
+    /// Premium animation for the sticker, if the sticker is premium.
+    pub premium_animation: Option<File>,
 
     /// For mask stickers, the position where the mask should be placed.
     pub mask_position: Option<MaskPosition>,